fix: stabilize variance dtype + widen ROILike (Copilot #170)
- normalize_transmission's background-ROI first-order variance term now casts the
  added contribution to transmission.variances.dtype, matching the float32
  discipline already in normalize_with_dark (issue #147). Defensive: the pure
  float32 path already stayed float32, but this guards mixed-dtype callers and
  keeps the two code paths consistent.
- ROILike now includes list[int], matching as_roi_bounds's docstring and the
  tests that pass a list (the alias previously only allowed ROI | tuple, causing
  avoidable type-checker noise).

424 tests pass; pre-commit clean.

Assisted-With: Claude Opus 4.8 <noreply@anthropic.com>

+++ src/neunorm/processing/normalizer.py
@@ -288,7 +288,8 @@ def normalize_transmission(  # noqa: C901
             co_term = co_var / (co * co)
             coeff_rel_var = co_term if coeff_rel_var is None else coeff_rel_var + co_term
         extra = sc.array(dims=list(transmission.dims), values=transmission.values**2) * coeff_rel_var
-        transmission.variances = transmission.variances + extra.values
+        # Keep the variance dtype stable (issue #147 float32 pipelines), matching normalize_with_dark.
+        transmission.variances = transmission.variances + extra.values.astype(transmission.variances.dtype, copy=False)
 
     logger.success("✓ Transmission normalized")
